backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
redis==5.0.1
aioredis==2.0.1
python-dotenv==1.0.0
email-validator==2.1.0
jinja2==3.1.2
aiosmtplib==3.0.1
